Move pylxd.container.LXDContainer to pylxd.deprecated.container
